+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Move all aerolab files and config paths to `~/.aerolab`.
 * When creating firewalls/security groups, by default these will be locked to the caller's IP address.
 * Allow to specify custom name for firewalls and security groups.
+* Make the `GCP` backend disks use the `NVME` interface for disks.
 
 #### 6.0.3
 * Handle bug in IP discovery on `docker` backend with custom networks.